refactor: move TxBuilder type in its own module
